design-spec: consolidate duplicate StyledFormMixin into a single shared mixin under main/mixins.py
