+++ account/forms.py
@@ -2,26 +2,9 @@ from django import forms
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth import get_user_model
 from django.utils.translation import gettext_lazy as _
+from main.mixins import StyledFormMixin
 
 User = get_user_model()
-
-
-class StyledFormMixin:
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        for field_name, field in self.fields.items():
-            if isinstance(field.widget, forms.CheckboxInput):
-                field.widget.attrs.update(
-                    {
-                        "class": "rounded-xs text-primary focus:ring-primary h-4 w-4 border-hairline bg-canvas"
-                    }
-                )
-            else:
-                field.widget.attrs.update(
-                    {
-                        "class": "rounded-sm border border-hairline bg-canvas text-ink px-4 py-3 w-full text-base focus:outline-none focus:ring-1 focus:ring-primary"
-                    }
-                )
 
 
 class RegistrationForm(StyledFormMixin, UserCreationForm):
